# Internalize Environment services and use Key as name