--- a/Jint.Tests.PublicInterface/RavenApiUsageTests.cs
+++ b/Jint.Tests.PublicInterface/RavenApiUsageTests.cs
@@ -34,9 +34,7 @@ public class RavenApiUsageTests
         var functionObject = new ScriptFunction(
             engine,
             functionExp,
-            engine.Advanced.CreateDeclarativeEnvironment(),
-            strict: false
-        );
+            strict: false);
 
         Assert.NotNull(functionObject);
     }