--- a/Jint/Native/Promise/JsPromise.cs
+++ b/Jint/Native/Promise/JsPromise.cs
@@ -1,4 +1,3 @@
-using Jint.Native.Function;
 using Jint.Native.Object;
 using Jint.Runtime;
 using Jint.Runtime.Descriptors;