--- a/Jint/Native/ShadowRealm/ShadowRealm.cs
+++ b/Jint/Native/ShadowRealm/ShadowRealm.cs
@@ -1,7 +1,6 @@
 ﻿using Esprima;
 using Esprima.Ast;
 using Esprima.Utils;
-using Jint.Native.Function;
 using Jint.Native.Object;
 using Jint.Native.Promise;
 using Jint.Runtime;