… Jint/Runtime/Environments/ModuleEnvironment.cs
… Jint/Runtime/Environments/ObjectEnvironment.cs
… Jint/Runtime/KnownKeys.cs
… Jint/Runtime/Modules/SyntheticModule.cs