Checked the usual suspects seems that `Environment` and related APIs can be hidden, allows using `Key` and only one to suffer a small hit is `ObjectEnvironment` which is used by `with` and its usage is discouraged.