@lahma 2024-01-13
## Jint.Benchmark.DromaeoBenchmark

| **Diff**|Method|Toolchain|FileName|Mean|Error|Allocated|
|------- |-------|-------|-------|-------:|-------|-------:|
| Old |Run|Default|dromaeo-3d-cube|22.317 ms|0.0417 ms|6284.06 KB|
| **New** |	| **Default** |	| **23.176 ms (+4%)** | **0.0239 ms** | **6284.27 KB (0%)** |
| Old |Run|Default|dromaeo-core-eval|4.356 ms|0.0225 ms|311 KB|
| **New** |	| **Default** |	| **4.063 ms (-7%)** | **0.0129 ms** | **311.47 KB (0%)** |
| Old |Run|Default|dromaeo-object-array|45.259 ms|0.4791 ms|100362.98 KB|
| **New** |	| **Default** |	| **45.056 ms (0%)** | **0.2089 ms** | **100362.98 KB (0%)** |
| Old |Run|Default|droma(...)egexp [21]|157.635 ms|1.7532 ms|161417.29 KB|
| **New** |	| **Default** |	| **161.615 ms (+3%)** | **1.2240 ms** | **165555.16 KB (+3%)** |
| Old |Run|Default|droma(...)tring [21]|334.605 ms|13.3655 ms|1321325.97 KB|
| **New** |	| **Default** |	| **318.525 ms (-5%)** | **12.3468 ms** | **1321588.91 KB (0%)** |
| Old |Run|Default|droma(...)ase64 [21]|51.122 ms|0.1005 ms|6044.92 KB|
| **New** |	| **Default** |	| **51.766 ms (+1%)** | **0.0552 ms** | **6044.57 KB (0%)** |
| Old |Run|Default|dromaeo-3d-cube|21.986 ms|0.0248 ms|5994.73 KB|
| **New** |	| **Default** |	| **21.807 ms (-1%)** | **0.0619 ms** | **5994.94 KB (0%)** |
| Old |Run|Default|dromaeo-core-eval|4.174 ms|0.0132 ms|298.44 KB|
| **New** |	| **Default** |	| **4.093 ms (-2%)** | **0.0076 ms** | **298.91 KB (0%)** |
| Old |Run|Default|dromaeo-object-array|44.652 ms|0.3315 ms|100324.02 KB|
| **New** |	| **Default** |	| **45.010 ms (+1%)** | **0.1634 ms** | **100324.02 KB (0%)** |
| Old |Run|Default|droma(...)egexp [21]|154.841 ms|1.2736 ms|172295.65 KB|
| **New** |	| **Default** |	| **160.887 ms (+4%)** | **1.5880 ms** | **165940.47 KB (-4%)** |
| Old |Run|Default|droma(...)tring [21]|334.037 ms|13.4039 ms|1321149 KB|
| **New** |	| **Default** |	| **366.460 ms (+10%)** | **14.3587 ms** | **1321478.8 KB (0%)** |
| Old |Run|Default|droma(...)ase64 [21]|57.075 ms|0.1819 ms|5958.77 KB|
| **New** |	| **Default** |	| **52.097 ms (-9%)** | **0.1542 ms** | **5958.35 KB (0%)** |